requirements.txt
# Core async and networking
asyncio
aiofiles==23.2.1
asyncssh==2.14.0

# Web frameworks
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
aiohttp==3.9.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
alembic==1.13.1

# Security and crypto
cryptography==41.0.7
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Threat Intelligence
geoip2==4.7.0
requests==2.31.0
aiohttp==3.9.1

# Machine Learning
scikit-learn==1.4.0
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2

# Logging and monitoring
python-json-logger==2.0.7
colorlog==6.8.0

# Reporting
jinja2==3.1.3
matplotlib==3.8.2
reportlab==4.0.8

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Alerts
python-multipart==0.0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
